update log format
