+++ examples/pom.xml
@@ -31,7 +31,7 @@
     <name>capa-sdk-examples</name>
 
     <properties>
-        <log4j.version>2.8.2</log4j.version>
+        <log4j.version>2.14.1</log4j.version>
         <logback.version>1.1.7</logback.version>
         <slf4j-api.version>1.7.32</slf4j-api.version>
     </properties>
@@ -61,17 +61,17 @@
             </exclusions>
         </dependency>
         <!-- slf4j -->
-        <dependency>
+       <!-- <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-slf4j-impl</artifactId>
             <version>${log4j.version}</version>
-        </dependency>
+        </dependency>-->
         <!-- logback -->
-        <!-- <dependency>
+         <dependency>
              <groupId>ch.qos.logback</groupId>
              <artifactId>logback-classic</artifactId>
              <version>${logback.version}</version>
-         </dependency>-->
+         </dependency>
         <!--lombok-->
         <dependency>
             <groupId>org.projectlombok</groupId>
