+++ examples/src/main/java/group/rxcloud/capa/examples/log/DemoLog.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package group.rxcloud.capa.examples.telemetry;
+package group.rxcloud.capa.examples.log;
 
 import lombok.extern.slf4j.Slf4j;
