+++ sdk-component/pom.xml
@@ -33,7 +33,7 @@
     <properties>
         <okhttp.version>4.9.1</okhttp.version>
         <kotlin-stdlib.version>1.4.10</kotlin-stdlib.version>
-        <log4j.version>2.8.2</log4j.version>
+        <log4j.version>2.14.1</log4j.version>
         <logback.version>1.1.7</logback.version>
         <guava-version>19.0</guava-version>
         <commons-lang3-version>3.11</commons-lang3-version>
