+++ sdk-component/src/main/java/group/rxcloud/capa/component/log/CapaLog4jAppenderAgent.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package group.rxcloud.capa.component.log.agent;
+package group.rxcloud.capa.component.log;
 
-import group.rxcloud.capa.component.log.enums.CapaLogLevel;
-import group.rxcloud.capa.component.log.manager.LogManager;
 import group.rxcloud.capa.infrastructure.CapaClassLoader;
 import org.apache.logging.log4j.core.Appender;
 import org.apache.logging.log4j.core.Filter;
@@ -30,7 +28,6 @@ import org.apache.logging.log4j.core.config.plugins.PluginElement;
 import org.apache.logging.log4j.core.config.plugins.PluginFactory;
 
 import java.io.Serializable;
-import java.util.Optional;
 
 /**
  * The abstract log4j appender. Extend this and provide your specific impl.
@@ -100,12 +97,13 @@ public class CapaLog4jAppenderAgent extends AbstractAppender {
 
     @Override
     public void append(LogEvent event) {
-        if (event != null && event.getLevel() != null) {
+        logAppender.append(event);
+       /* if (event != null && event.getLevel() != null) {
             Optional<CapaLogLevel> capaLogLevel = CapaLogLevel.toCapaLogLevel(event.getLevel().name());
             if (capaLogLevel.isPresent() && LogManager.whetherLogsCanBeOutput(capaLogLevel.get())) {
                 logAppender.appendLog(event);
             }
-        }
+        }*/
     }
 
     /**
@@ -118,6 +116,6 @@ public class CapaLog4jAppenderAgent extends AbstractAppender {
          *
          * @param event The log event.
          */
-        void appendLog(LogEvent event);
+        void append(LogEvent event);
     }
 }
