+++ sdk-component/src/test/java/group/rxcloud/capa/component/log/TestCapaLogbackAppender.java
@@ -14,17 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package group.rxcloud.capa.component.log.agent;
+package group.rxcloud.capa.component.log;
 
-import org.apache.logging.log4j.core.LogEvent;
+import ch.qos.logback.classic.spi.ILoggingEvent;
 
 /**
- * The capa log4j appender used in tests only.
+ * The capa logback appender used in tests only.
  */
-public class TestCapaLog4jAppender implements CapaLog4jAppenderAgent.CapaLog4jAppender {
+public class TestCapaLogbackAppender implements CapaLogbackAppenderAgent.CapaLogbackAppender {
+
+    public TestCapaLogbackAppender() {
+    }
 
     @Override
-    public void appendLog(LogEvent event) {
-        System.out.println("test log log4j and content is " + event.getMessage().getFormattedMessage());
+    public void append(ILoggingEvent event) {
+        System.out.println("test logback log and content is " + event.getFormattedMessage());
     }
 }
+
