+++ sdk-component/src/test/resources/capa-component-log.properties
@@ -1,2 +1,2 @@
-group.rxcloud.capa.component.log.agent.CapaLog4jAppenderAgent$CapaLog4jAppender=group.rxcloud.capa.component.log.agent.TestCapaLog4jAppender
-group.rxcloud.capa.component.log.agent.CapaLogbackAppenderAgent$CapaLogbackAppender=group.rxcloud.capa.component.log.agent.TestCapaLogbackAppender
+group.rxcloud.capa.component.log.CapaLog4jAppenderAgent$CapaLog4jAppender=group.rxcloud.capa.component.log.TestCapaLog4jAppender
+group.rxcloud.capa.component.log.CapaLogbackAppenderAgent$CapaLogbackAppender=group.rxcloud.capa.component.log.TestCapaLogbackAppender
