+++ sdk-spi-demo/pom.xml
@@ -31,7 +31,7 @@
     <name>capa-sdk-spi-demo</name>
 
     <properties>
-        <log4j.version>2.8.2</log4j.version>
+        <log4j.version>2.14.1</log4j.version>
         <logback.version>1.1.7</logback.version>
     </properties>
